+++ metatensor-torch/src/atomistic/model.cpp
@@ -844,7 +844,7 @@ ModelMetadata metatensor_torch::read_model_metadata(std::string path) {
     if (!reader.hasRecord("extra/model-metadata")) {
         C10_THROW_ERROR(ValueError,
             "could not find model metadata in file at '" + path +
-            "', did you export your model with metatensor-torch >=0.6?"
+            "', did you export your model with metatensor-torch >=0.5.4?"
         );
     }